Switched to BasicColumn<int64_t> due to int soon-to-be deprecated

+++ test/test_column_basic.cpp
@@ -41,8 +41,8 @@ using namespace realm;
 TEST(ColumnBasic_LowerUpperBound)
 {
     // Create column with sorted members
-    ref_type ref = BasicColumn<int>::create(Allocator::get_default());
-    BasicColumn<int> col(Allocator::get_default(), ref);
+    ref_type ref = BasicColumn<int64_t>::create(Allocator::get_default());
+    BasicColumn<int64_t> col(Allocator::get_default(), ref);
 
     col.add(5);
     for (int i = 5; i < 100; i += 5)
@@ -78,8 +78,8 @@ TEST(ColumnBasic_LowerUpperBound)
 
 TEST(ColumnBasic_NullOperations)
 {
-    ref_type ref = BasicColumn<int>::create(Allocator::get_default());
-    BasicColumn<int> c(Allocator::get_default(), ref);
+    ref_type ref = BasicColumn<int64_t>::create(Allocator::get_default());
+    BasicColumn<int64_t> c(Allocator::get_default(), ref);
 
     CHECK(!c.is_nullable());
     CHECK(!c.is_null(0));
@@ -89,8 +89,8 @@ TEST(ColumnBasic_NullOperations)
 
 TEST(ColumnBasic_NullErrorHandling)
 {
-    ref_type ref = BasicColumn<int>::create(Allocator::get_default());
-    BasicColumn<int> c(Allocator::get_default(), ref);
+    ref_type ref = BasicColumn<int64_t>::create(Allocator::get_default());
+    BasicColumn<int64_t> c(Allocator::get_default(), ref);
 
     CHECK_LOGIC_ERROR(c.set_null(0), LogicError::column_not_nullable);
